protected user route and client side redirect if not logged in for create blog an update blog pages

+++ src/app/api/blog/[id]/route.ts
@@ -39,17 +39,37 @@ export async function GET(req: NextRequest, { params }: RouteParams) {
   }
 }
 
+//PROTECTED ROUTE
 export async function DELETE(req: NextRequest, { params }: RouteParams) {
   try {
+    //access the request header for user ID
+    const userID = req.headers.get("X-user-ID");
+    if (!userID) {
+      return NextResponse.json({ error: "not authorized" }, { status: 401 });
+    }
+
+    //get blog id
     const { id } = await params;
 
     if (!id) {
       return NextResponse.json({ error: "invalid URL" }, { status: 400 });
     }
 
-    const blog = await prisma.blog.delete({ where: { id: Number(id) } });
-
+    //is user creater of the blog?
+    const blog = await prisma.blog.findUnique({ where: { id: +id } });
     if (!blog) {
+      return NextResponse.json(
+        { message: "could not find blog" },
+        { status: 404 }
+      );
+    }
+    if (+userID !== blog.authorID) {
+      return NextResponse.json({ message: "not authorized" }, { status: 401 });
+    }
+
+    const deletedBlog = await prisma.blog.delete({ where: { id: Number(id) } });
+
+    if (!deletedBlog) {
       return NextResponse.json({ error: "blog not found" }, { status: 404 });
     }
 
@@ -64,6 +84,7 @@ export async function DELETE(req: NextRequest, { params }: RouteParams) {
   }
 }
 
+//PROTECTED ROUTE
 export async function PATCH(req: NextRequest, { params }: RouteParams) {
   try {
     //access the request header for user ID
